requirements.txt
slack-bolt
python-dotenv
notion-client 